// src/repository/user.repository.d.ts
declare class UserRepository {
    getUsers(): Promise<{
        id: string;
        name: string;
        surname: string;
        age: number;
        email: string;
        password: string;
        phone: string | null;
        gender: import("../../prisma/src/generated/prisma/enums.ts").GenderEnum | null;
        photo: string | null;
        role_id: number;
// ...
    }[]>;
// ...
}
export declare const userRepository: UserRepository;